refactored collapse controlling targets into target-controls.vue
