+++ app/frontend/vue/home/analysis/apexChartHelpers.js
